Add panviewport example
The idea is to have the best possible example to handle a viewport
with different devices on all platforms. It needs more testing and we
will try to document each case and device how it's handled and what
information is reported (example of devices are mouse, mouse wheels,
touchpads, touchpad gestures, wacom tablets/stylus, surface screen and
pen, etc.).