Fix SkiaSurface::drawRect() size when we stroke a rectangle
Skia strokes a rectangle with width=height=1 as a 2x2 rectangle
instead of 1x1 (this doesn't happen when we use the kFill_Style). But
we were painting (stroke) rectangles of 1x1 as a pixel. So here we fix
this regression introduced in dd9f578f924f1f8901920b3999350bfe2aaee29a,
related to https://github.com/aseprite/aseprite/issues/2294