Don't use wintab32.dll if we know that WTOpen() crashed before
We can create a temporal file if calling WTOpen() crashes (sometimes,
a buggy driver or wintab32.dll can crash our program). In this way we
can run a second time the program disabling the Wintab API
automatically.

This is not "an optimal" solution, but it's good enough given the
alternative, e.g. we cannot execute our app anymore until we restart
the computer/the driver is back to a normal state.