[win] Fix issue with Snipping Tool
Using the Snipping Tool sent the window to the back, losing the
focus/active window status. It looks like we cannot change the
WS_VISIBLE style at this point. So to avoid drawing the frame on a
WM_NCACTIVE message, we can temporarily remove the WS_*SCROLL and
WS_THICKFRAME styles.