Use SkPixmap inside SkiaSurface to render rectangles & avoid copying SkBitmap pixels
The Skia API is evolving in such a way to make it more GPU-friendly,
i.e. encouraging the usage of immutable SkBitmaps/SkImages. We are not
yet ready to make this change as we modify (re-use) SkiaSurfaces in
several places (mainly when we render our sprite editor surface).

In the new Skia m96 (and some previous versions),
SkImage::MakeFromBitmap() creates a copy of the pixels from mutable
SkBitmaps. So to avoid copying those pixels of each mutable SkBitmap,
we can use the alternative SkImage::MakeFromRaster(const SkPixmap&),
using the internal SkPixmap inside the SkBitmap.

In a future we should provide a new API to create different kind of
surfaces (immutable, raster, render target, etc.) depending on what
kind of usage we are going to do with the surface.