[win] Ignore random WM_MOUSEMOVE messages in the same mouse position
This solves a lot of issues when multiple pointer devices (e.g. a
mouse and a pen) are plugged in and we receive random WM_MOUSEMOVE
messages from the mouse position even when the device is not moved at
all (so these messages are usually mixed up with the pen messages from
WM_POINTERUPDATE).

Fixes https://github.com/aseprite/aseprite/issues/3433