Add native file dialog for Linux using zenity
We've removed the GTK+ impl and LAF_OS_WITH_GTK flag, as we don't have
plans to link with GTK+, and replace it with a new implementation that
run Zenity to show the dialog. It's not perfect, there are several
issues to improve the UX, but it's good enough to offer a native file
dialog on Linux.

Related to:
https://github.com/aseprite/aseprite/issues/1953
https://github.com/aseprite/aseprite/issues/2354