[win] Fix the repeat field of KeyDown event
0-15 bits of the LPARAM argument aren't the number of times that a
WM_KEYDOWN message was auto-repeated by just pressing a key down. It's
the number of enqueued messages that are not yet processed.

Anyway we can use the bit 30 of the LPARAM to know if the WM_KEYDOWN
message was generated when the key was previously up (the user pressed
it for the first time) or down (the user keeps pressing the key so
more messages are received from Windows).

Related to:
https://github.com/aseprite/aseprite/issues/3684
https://github.com/aseprite/aseprite/pull/4063
https://github.com/aseprite/laf/pull/69
https://github.com/aseprite/laf/pull/70