We can upgrade readers to writers when the current thread already has a writer lock
This is a missing condition in RWLock::canWriteLockFromRead() that is
possible in RWLock::upgradeToWrite() when the current thread already
has a writer lock. This case is handled as a re-entrant
lock (LockResult::Reentrant) and means that the thread can upgrade any
reader to writer anyway.

Completes: 157533d1600d424e089be859148aa71a169111f2
Related to: https://github.com/aseprite/aseprite/issues/4310