Add new TextBlob class and text_shape example
Now we have TextBlob::Make() and TextBlob::MakeWithShaper() to
cache/store the shaped layout of a set of glyphs to be rendered with
draw_text(..., TextBlobRef, ...). We can also iterate clusters with
RunHandler.

There is still some work to be done for RTL text and composed
chars. And we should try to replace the old DrawTextDelegate with the
new TextBlob::RunHandler in the future.