Make listscreens example work even when backend=none
We've refactored the System hierarchy so we can make an instance of
SystemWin/OSX/X11 classes even when LAF_BACKEND=none. We've made
System::makeWin/OSX/X11() functions available as they might be enough
in certain cases where the whole Skia is not needed.

Some important changes in this patch:
* os::instance() is deprecated, replaced with os::System::instance()
* os::make_instance() is deprecated, replaced with os::System::make()
* os::System::instance() returns an os::SystemRef