Add DrawTextDelegate support with Skia native fonts
Now we can use the draw_text() that receives a DrawTextDelegate with
native fonts. Anyway we should remove this function in the
future (it's marked as deprecated now) and use TextBlob directly.

With this patch the complextextlayout example is working again, we can
hover chars with the mouse to highlight them (and now works with
emojis).