Fix SkiaSurface::getData() when the SkiaSurface is created from a SkSurface_Raster
The SkSurface_Raster is a wrapper for a SkBitmap, we should be able to
get the underlying SkBitmap pixels to paint in the SkSurface_Raster
directly (e.g. used in draw_text() for FreeTypeFont).

This requires a special patch in Skia library that includes a new
member function SkSurface_Raster::skBitmap().